bug 1764570: rename librevision to libdockerflow
This reduces some ambiguity around "revision" and "version" plus it
gives us a good place to add other dockerflow things later on.
